chore(VMware): Modify section of instance-id in the customization config (#5356)
For the traditional guest os customization, there is no instance-id provided in the customization
configuration. While since the next vSphere release, instance-id will be provided and included
in the traditional guest os customization configuration for the first time, the format is
"MISC|INSTANCE-ID".
This change is correcting the config parser and related test.