Improve schema validator warning messages
During initialization, schema validator may throw warning messages if it finds
the provided configuration does not agree with the schema definition:

schema.py[WARNING]: Invalid cloud-config provided

In reality, even if the configuration does not agree with the schema, it may
sometimes still work. Improve the error message to make it more factual instead
of broadly claiming that the config is broken.

fixes: GH-5399
Signed-off-by: Ani Sinha <anisinha@redhat.com>