doc: update examples to reflect alternative ways to provide `sudo` option (#5418)
For creating users and groups, it is possible to pass a `sudo` option to the
config file that accepts a sudo rule. The option can be a sudo rule string,
a list of sudo rule strings or `False` to explicitly deny sudo usage. Update
examples to show how a list of strings can be used with `sudo` option.

Signed-off-by: Ani Sinha <anisinha@redhat.com>