feat(snap): avoid refresh on package_upgrade: true and refresh.hold
When snap refresh.hold is set to forever. An admin is saying they do
not want generic automated refreshes of snaps performed by default.

This should be an indicator to cloud-init to avoid calling snap refresh
on such systems due to a `package_upgrade: true` present in user-data.

For network-limited environments with images which contain the snap
package manager bu don't want to wait and timeout on snap refresh,
the following user-data can be provided to still allow for
package_upgrade: true, but avoid a 20-30 second wait on snaps being
unable to access certain snap URLs.

   #cloud-config
   package_upgrade: true
   snap:
     commands:
       00: snap refresh --hold

cloud-init now interrogates the state refresh.hold by calling
  snap get system refresh.hold

If snap refresh --hold was called in that environment with a 'forever'
value, then cloud-init will skip calling refresh and log the reason for
that behavior.

Fixes: GH-5290