doc: update qemu command line (correctness)
From the qemu docs we know that the old arguments are not
encouraged:
  "Older options like -hda are essentially macros which expand into -drive
   options for various drive interfaces. The original forms bake in a lot of
   assumptions from the days when QEMU was emulating a legacy PC, they are
   not recommended for modern configurations."

Also by not specifying the format we get warnings like:
 WARNING: Image format was not specified for 'seed.img' and probing guessed raw.

These warnings might one day get more aggressive and so might be the use
of -hda. Instead let us avoid the current and the potential future by
using the new format which is a bit more verbose, but also more correct.