Revert "fix(vmware): Set IPv6 to dhcp when there is no IPv6 addr (#5471)"
This reverts commit 2b6fe6403db769de14f7c7b7e4aa65f5bea8f3e0.

When there is no IPv6 set to dhcp explicitly, NetworkManager keyfile defaults
to method=auto, may-fail=false.
When there is Ipv6 set to dhcp explictily, NetworkManager keyfile will be
set to method=auto, may-fail=true.
The default settings are what we want, so revert the previous change to keep
IPv6 not set explicitly.