Models non-mergeable data as an #include file
If either agent or user/Landscape data is not a #cloud-config, then we
cannot apply the merge rules of this datasource, but rather we need to
pass them along so cloud-init internals can handle them.

With the current implementation we attempt to pass them as a list of
bytes. That has lots of limitations and breaks cloud-init internal
handlers in many ways. For example a user data YAML file containing a
syntax error would trigger a nasty traceback and it would be hard to
find in the logs that the error was due a ill-formed YAML (thanks jibel
for showing me that).

Investigating edge cases like that I also realized we couldn't support
jinja templates of any sort with that implementation.

So, if one of the merge candidates is not cloud-config, the best way to
feed cloud-init with them is by pretending the user wrote an #include
file. That should allow all cloud-init features to just work.

Previously we didn't check if the YAML we loaded was actually
cloud-config. It could be a jinja template of a YAML cloud-config, in
which case merging wouldn't make sense. Thus we now explicitely check
before loading the YAML as a dict. A jinja template of a cloud-config
for example goes raw into cloud-init, which is exactly what we want.
UserDataProcessor's will handle the template rendering and
post-processing just fine.