fix(scaleway): Don't override IPv6 routes when IPv4 is not primary
If an instance in _routed-ip_ mode uses both IPv4 **and** IPv6, but the
IPv6 is listed (attached) first, the default IPv6 route may be absent
from the final network configuration for systems that do not accept
Router Advertisements.

This changes fixes the problem by making sure that routes are appended,
not overwritten.