fix: lxd do not check for thinpool kernel module
Rather than checking for kernel module file before running lxd, just
fall back to disabling thinpool and retrying with thinpool disabled
when cloud-init fails and it is using lvm as backing storage.

On some Ubuntu releases, the kernel module is compressed which causes
this check to fail and this warning to erroneously be logged.