fix: lxd do not check for thinpool kernel module (#5709)
Rather than checking for kernel module file before running lxd, just
fall back to disabling thinpool and retrying with thinpool disabled
when subp() fails and it is using lvm as backing storage.

On some Ubuntu releases, the kernel module is compressed which causes
this check to fail and this warning to erroneously be logged. Making
this logic only run in the failure path and avoiding file matches on
kernel driver files should be less error prone.

BREAKING_CHANGE: On some Ubuntu series this changes behavior causing
lxd to now use lxd thinpool, which is the current lxd default.