test: Fix test_log_message_on_missing_version_file
This test has some questionable history that has been removed. The
main purpose of the test is to ensure that we get a log that the
python-version file is written if there's no pre-existing
python-version file but there is pre-existing pickle cache.

The invalid pickle at the beginning of the test complicated this test.
A normal boot will write a valid cache, so we should get no pickle
errors on reboot.