fix: Add new func param to test case call point
The generate_sources_list function signature now has a parameter, "keys". Rather
than using a default value, the call points have been updated to reflect the
change. In the case of the call in test_cc_apt_configure, that meant passing an
empty dictionary to the function in the appropriate test case