feat: Custom keys for apt archives (#5828)
Users with local Ubuntu archive mirrors and Landscape instances have
been unable to specify the corresponding gpg keys. This resulted in
errors such as "NO_PUBKEY" on commands such as "apt update".

This commit adds the functionality to supply keys alongside primary
and security mirror declarations. The key can either be defined using
the "key" mapping, or via a keyid and (optionally) a keyserver. Using
either approach, when a key is supplied alongside the primary or
security mirror it is now added to /etc/apt/trusted.gpg.d/ as
primary.gpg or security.gpg accordingly and the Signed-By field in the
deb822 templates are appropriately populated with this value.

If no primary key is supplied, it defaults to the ubuntu archive keyring. If no
security key is supplied, it falls back on the primary key (to match the
behaviour of the security URI falling back on the primary URI), and in turn
falls back on the ubuntu archive keyring if that is not defined.

Fixes GH-5473