fix: Use /usr/lib/ rather than /lib in packaging code (#5970)
Support for unmerged /usr was removed from systemd a while ago:
https://lists.freedesktop.org/archives/systemd-devel/2022-September/048352.html

Furthermore, Debian tooling will now throw lintian errors if files
continue to reference /lib. Since it is no longer supported in upstream
systemd, there is no reason for it to be supported in upstream
cloud-init code that uses systemd.