net/sysconfig: do not remove all existing settings of /etc/sysconfig/network
In some distros, /etc/sysconfig/network may have important configurations that
are necessary for the instance to come up. For example, centos based distros
write NOZEROCONF=yes in /etc/sysconfig/network for some instances that require
zeroconf to be disabled. Removing these customizations would prevent the
instance to come up. So leave the customizations in /etc/sysconfig/network
intact except those that we are interested in.

Fixes GH-5990
Signed-off-by: Ani Sinha <anisinha@redhat.com>