fix: NM reload and bring up individual network conns (#6073)
Reloading the network manager service is equivalent to "nmcli reload" and this
command only reloads the global .conf files and DNS config, not connections.
This means changes to connection files will not take effect. For those to take
effect, we need "nmcli conn load/reload" and then "nmcli conn up". Thus,
reloading network manager as well as reloading the connections are required to
cover all cases.

Also see https://github.com/canonical/cloud-init/issues/5512#issuecomment-2298371744

While at it, rename "reload-or-try-restart" -> "try-reload-or-restart" since
the former is legacy and the later is the officially documented sub-command.

Fixes: GH-6064
Fixes: bde913ae242 ("fix(NetworkManager): Fix network activator")

Signed-off-by: Ani Sinha <anisinha@redhat.com>