Fix: Add Ephemeral Network for CloudStackLocal DS
This commit adds a new local datasource, CloudStackLocal, which supports
ephemeral networking and parallels the structure of other cloud provider
datasources. In particular this functionality is enabled by performing DHCP
setup contingently on the perform_dhcp_setup parameter. Each detected NIC is
presented sequentially as a candidate for networking and performs DHCP
discovery. If a lease is not obtained, a subsequent NIC is tried. Otherwise
upon success, we try to reach the IMDS. If this were to fail for some reason,
we continue to the next NIC. Upon successful data retrieval from the IMDS we
return. The communication with the IMDS has been wrapped in a new function:
_crawl_metadata for improved modularity.

The motivation for this patch is outlined in GH issue #6143, which highlighted
the inability to use NetworkManager as a renderer in a CloudStack environment.
Symptomatically, cloud-init was unable to discover the IMDS by DNS or reach it
as no network had yet been created, and forcing cloud-init to wait on the
network resulted in a deadlock where NetworkManager required dbus.sock but
could not acquire it until cloud-init.service finished - which could not happen
without the network. The ephemeral network in this commit ensures that
networking is available at the time that we reach out to the IMDS - ensuring
it can be discovered dynamically and avoiding the deadlock.