feat(vmware): Support network events (#6063)
This patch updates the DataSource for VMware to support network
reconfiguration when the BOOT, BOOT_NEW_INSTANCE, and HOTPLUG
events are received.

Previously the datasource could only reconfigure the network if
a new instance ID was detected. However, due to features like
backup/restore, migrating VMs, etc., it was determined that it is
valuable to support reconfiguring the network without changing the
instance ID. This is because changing the instance ID also means
running the per-instance configuration modules, such as regenerating
the system's SSH host keys, which could lock out automation.