fix(oracle): properly detect ipv6 only for private ULA addresses (#6123)
When a VNIC is attached to a private subnet and has a ULA address,
the VNIC data given by the IMDS does not contain the
"ipv6SubnetCidrBlock" entry and instead has a "subnetCidrBlock" entry
set to null. To properly detect if we should configure a VNIC as ipv6
single stack, the key "ipv6VirtualRouterIp" must be present and the key
"privateIp" must not be present. Previously, the key
"ipv6SubnetCidrBlock" was assumed to be present on all VNICs configured
for ipv6 single stack or dual stack, but that is not the case when using
a private subnet with ULA addresses.