chore: reorder iface filters & log on inherited MAC (#6140)
This filter was originally introduced in bf7723e to prevent bonds and
vlans being included in `get_interfaces`. However, specific filtering
for vlans was implemented in ebc9ecb and the same was added for bonds
in e5f5421.

As of kernel commit cb6cf08, some USB NICs which use passthrough MAC
addresses will also report their MAC address as stolen. MAC
passthrough makes it possible for more than one physical interface to
use the same MAC, so provide a log to make it clear what
is going on.