Wait for systemc nics to initialize and don't break cloud-init
During boot process cloud-init local stage is very early and system nics
are not always initialized yet. If a cloud-init is provided with a system
nic not yet present, the whole cloud-init process will fail and not even
a debug user will get created. Change the behaviour to increase the number
of retries when waiting for system nics to get initialized and if still
not possible use a generated nic name instead of failing completely.

Signed-off-by: Matthias Haag <matthias.haag@uhurutec.com>