[Feat]: Print Network Device Table After DBUS
When using NetworkManager as the network render, the network device table that
is printed to the console does not contain the final state of IP addresses for
network devices. While this is not inherently a bug as the table is intended to
be debug information for a specific class of boot failures rather than
information level for end-users, many cloud-init consumers assess the success of
the network setup stage against this table. The reason the table does not show
the final results for NetworkManager is that at the point in time that it is
logged, NetworkManager has not setup interfaces yet. It does this at a later
point due to a dependency on DBUS.

This commit prints a second table at the start of the final stage when
NetworkManager will have setup the interfaces. Moreover, to prevent confusion
surrounding the presence of two tables, each table is preceeded by a line that
indicates whether the following table represents the network state before or
after DBUS.

This addresses the discussion in GH issue #6161