[Fix]: Refactor Unit Tests and CloudStack's init
This commit changes DataSourceCloudStack's initialization, moving the
get_vr_address call into the _get_data method within a network context as it
makes little sense to call this function without the proper network
configuration established. The network context is the ephemeral network for
DataSourceCloudStackLocal, and util.nullcontext when not performing dhcp setup
(DataSourceCloudStack). The DataSource is also updated to include a ds_detect
method which calls the new is_platform_viable method to determine if we are
running this datasource.

This commit also refactors the unit tests to use pytest.fixtures, and avoid
inheritance from the CiTestCase class.