Fix: Add Ephemeral Network for CloudStackLocal DS (#6144)
Add a new local datasource, CloudStackLocal, which supports
ephemeral networking and parallels the structure of other cloud provider
datasources. In particular this functionality is enabled by performing DHCP
setup contingently on the perform_dhcp_setup parameter. The datasource
will rely on the primary of fallback NIC as the candidate for networking and
will perform DHCP discovery. Upon success, we try to reach the IMDS.
Upon successful data retrieval from the IMDS we return.
The communication with the IMDS has been wrapped in a new function:
_crawl_metadata for improved modularity.

Additionally platform detection logic is_platform_viable mirrors ds-identify
logic for datasource validation to assert that the CloudStack datasource is
detected via DMI server-product-name values from the platform. This validation
is surfaced by DatasourceCloudStack._ds_detect which is called before _get_data
is allowed to run on a given system. Platforms which do not surface DMI
information advertising CloudStack datasource will not be able to run this
datasource.

The motivation for this patch is outlined in GH issue #6143, which highlighted
the inability to use NetworkManager as a renderer in a CloudStack environment.
Symptomatically, cloud-init was unable to discover the IMDS by DNS or reach it
as no network had yet been created, and forcing cloud-init to wait on the
network resulted in a deadlock where NetworkManager required dbus.sock but
could not acquire it until cloud-init.service finished - which could not happen
without the network. The ephemeral network in this commit ensures that
networking is available at the time that we reach out to the IMDS - ensuring
it can be discovered dynamically and avoiding the deadlock.

Fixes GH-6143