chore!: switch to meson.build due to PEP517 setuptools data_files depr
setuptools deprecated use of data_files for non-package external data
files or system files for some time[1],[2] as part of PEP517[3] work.

This PEP517 incompatiblity prompts some distributions to drop support
for cloud-init in their operating system[4].

Rather than pursue the use of setup.py via setuptools further, shift
to meson which better supports downstream operating system-level
packaging of ancillary distribution of man pages,
config files, udev, openrc or systemd scripts which are intended to
live outside of a pure python distribution.

Since cloudinit package is primarily intended to be delivered as system
library and utility for initial system setup during first boot,
it is out of the scope of cloud-init to support local wheel or bdist
installations via pypi in custom locations outside of system
packages as we want only one system-package installation of
cloud-init in any environment.

As such, cloud-init's use of python setuptools is not necessary.
A move to meson gives a bit more flexibility to handle parameterized
delivery of the supplemental config, init scripts, man pages an
tooling which can be extended for cross-platform build support.

This commit creates an initial draft of the work to migrate cloud-init
backend to use meson instead.

- Drop setup.py and replace that functionality with meson.build files
  which perform the same setup, template rendering and installs
  previously done by setuptools.
- Add test targets and dependency install targets to meson.build
- Set pyproject.toml to use meson build backend
- Drop now unused setup_utils and moving pep400 func to test_tools.py
- Add make clean target removal of meson buildir

What is missing for this commit before merging:
- cloud-init documentation page in doc/rtd explaining how to leverage
  meson for
- a commit addressing packaging/ templates for debian, redhat and suse to
  represent applicable changes for RPMs/DEBs
- a separate downstream example pull request for ubuntu enabling meson
  deb package builds

References:
- [1] https://setuptools.pypa.io/en/latest/references/keywords.html
- [2] https://setuptools.pypa.io/en/latest/userguide/datafiles.html#non-package-data-files
- [3] https://peps.python.org/pep-0517/

Fixes: GH-5027