Docs: Network Device Table and Limitations
This commit adds documentation to highlight that the network device table which
is printed to the terminal during boot is intended to be debug-level output
rather than user-level information. Additionally, it explains potential
justifications for unexpected output and references the debugging docs as a
method of further investigation.

This addresses the discussion in GH issue #6161

Fixes: GH-6161