eni.py: ensure that a "dns" entry is not rendered in interfaces file (#6253)
Problem: when either a v1 or v2 network-config that contain DNS-related
entries is rendered to eni format then the resultant interfaces file
generated contains a "dns {...}" entry that should not be present, in
addition to the expected "dns-nameservers" and/or "dns-search" entries.

Fix: change the eni rendering code to ignore any "dns" section present
in cloud-init's internal JSON network config representation.