docs: provide example3 for PAM and ssh_pwauth behavior (#27)
On certain distributions, default behavior of the distribution
may be to set sshd_config `KbdInteractiveAuthentication yes` which
allows password authetication in environments where PAM modules are
used for authentication.

If cloud-config user-data contains `ssh_pwauth: false` this only
informs cloud-init to disable the sshd_config setting
`PasswordAuthentication no` which may not be sufficient if the target
environment is both using PAM for authentication and has a default
of `KbdInteractiveAuthentication yes`.

Provide documentation in module examples for providing user-data
in such scenarios to also disable KbdInteractiveAuthentication if
warranted.