fix: allow downstreams to strip invalid MTU (#6246)
mtu: 0 is an invalid value within neptlan configuration, but cloud-init
allowed it in previous releases. Since this also applies to v2
passthrough, we need strip the line outside of normal parsing.
This commit adds a flag to allow cloud-init to strip it from
downstreams to keep backwards compatibility.

Fixes GH-6239