feat(oracle): set keep_configuration to true for iscsi instances
This is necessary for Oracle Baremetal instances on IPv6-only networks
which rely on the iscsi connection. Without this, after shutting down
the instance, the instance is not recoverable.

This setting will be automatically set for all iscsi instances via the
Oracle Datasource based on metadata/config files left behind by
initramfs during boot on ISCSI instances.

PR: #6097