clean: do not attempt to find datasource when cache has been cleaned (#6325)
When `cloud-init clean` is run with or without additional options but without
`-s`, it removes everything except the seed directory from `/var/lib/cloud`.
When `-s` is passed with or without other options, it removes the seed directory
as well. A subsequent invocation of `cloud-init clean -c datasource` will fail
to find the current datasource as the cache has been cleaned. Therefore, when
cache is not present, do not try to find the datasource.

Fixes: 34cb46e20f5d3f ("feat: add callback for datasources to clean config changes (#6100)")

Signed-off-by: Ani Sinha <anisinha@redhat.com>