feat(password): support type=RANDOM_FRIENDLY
In GH PR #5815, the previous random password schema was enhanced to
support generation of stronger passwords.  These passwords can however
be challenging for a human to read due to excessive punctuation and
ambiguous characters. (quick - lI1 - which character is which?  Your
results here will vary depending on the display font)

Add a new password type RANDOM_FRIENDLY that generates passwords like
the pre-PR5815 scheme, with no punctuation and excluding the same set of
ambiguous characters.

LP: #2118989