net_convert.py: make some import failures not generate an error (#6399)
If cloud-init DataSources are separately packaged by a distro then it
is possible that running "cloud-init devel net-convert -h" (and trying
to convert between some network formats) can fail with errors if
Azure, Openstack, or VMware cloud-init code is not present.

I observed these errors when running the above command on an Alpine
system where cloud-init's various DataSources are separately packaged
and only the NoCloud DataSource was installed.

This PR prevents import failures of Azure, Openstack, and VMware
specific code from causing such errors. Instead the accepted list of
inputs ("-k") is dynamically adapted based on the presence/absence of
each of the Azure/Openstack/VMware related code.

NOTE: I realise that cloud-init does not currently officially support
the packaging of DataSources code separately. However I believe that
Ubuntu is currently in the process of developing/testing such separate
DataSources packaging and so this PR would be relevant to Ubuntu also.