fix(net-convert): respect output directory with netplan
Respect --directory when --output-kind is netplan.

Previously, if the netplan python API was present, cloud-init was
unconditionally ignoring the directory target and writing to
/etc/netplan/50-cloud-init.yaml or a similar location using the netplan
API. That writing operation only succeeded if the user had write
permission under /etc/netplan.

In the no-permission case, cloud-init falls back to manually writing to
$target + etc/netplan/50-cloud-init.yaml and that's why tests didn't
catch this issue.

Fixes: GH-6141