rhel: Do not override changes in disable-sshd-keygen-if-cloud-init-active.conf
When customer has local changes in disable-sshd-keygen-if-cloud-init-active.conf
and they try to upgrade cloud-init, the local changes are lost and the config
file reverts back to original content. This disrupts the customer as it breaks
their deployments and the config changes needs to be manually applied again.
The customer should be able to overwrite the config file with any changes that
work for their environment or work around some issues. Those changes should be
retained by rpm upgrade process. Therefore, in the spec file, mark the config
file as "noreplace" so that changes are kept.

Signed-off-by: Ani Sinha <anisinha@redhat.com>